Move test to another suite
